requirements-railway.txt
crewai
langchain
langchain-anthropic
langchain-groq
fastapi==0.111.0
uvicorn==0.29.0
celery==5.3.6
redis==4.6.0
apscheduler==3.10.4
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv
httpx==0.27.0
python-telegram-bot==21.1.1
pillow==10.3.0
boto3==1.34.102
sentry-sdk==2.1.1
loguru==0.7.2
langchain-google-genai==2.0.10
opencv-python-headless==4.9.0.80
imageio==2.34.0
imageio-ffmpeg==0.4.9
